+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - main:  Fixed handling command line arguments from SonarQube scan finding.
 
 ### Changed
+- run_program:  Setup and configured email.
 - main:  Added -e and -s options for email capability.
 - main:  Added -y option and program lock instance.
 - Documentation updates.